Add configurations to fill db.
Major changes:
- Add AddUserConfiguration, IdentityRoleConfiguration, IdentityUserRoleConfiguration;
- Apply configuration at OnModelCreating.
