+++ src/services/Identity/Identity.Infrastructure/DbContext/AuthDbContext.cs
@@ -1,0 +1,20 @@
+﻿using System.Reflection;
+using Identity.Domain.Entities;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore;
+
+namespace Identity.Infrastructure.DbContext
+{
+    public class AuthDbContext : IdentityDbContext<AppUser, IdentityRole<int>, int>
+    {
+        public AuthDbContext(DbContextOptions<AuthDbContext> options)
+            : base(options) { }
+
+        protected override void OnModelCreating(ModelBuilder builder)
+        {
+            base.OnModelCreating(builder);
+            builder.ApplyConfigurationsFromAssembly(Assembly.GetExecutingAssembly());
+        }
+    }
+}
