+++ src/services/Identity/Identity.Infrastructure/DbContext/DbConfiguration/AppUserConfiguration.cs
@@ -1,0 +1,24 @@
+using Identity.Domain.Entities;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+namespace Identity.Infrastructure.DbContext.DbConfiguration;
+
+public class AppUserConfiguration : IEntityTypeConfiguration<AppUser>
+{
+    public void Configure(EntityTypeBuilder<AppUser> builder)
+    {
+        var hasher = new PasswordHasher<AppUser>();
+        var admin = new AppUser()
+        {
+            Id = 1,
+            UserName = "admin",
+            NormalizedUserName = "ADMIN",
+            SecurityStamp = Guid.NewGuid().ToString(),
+            PasswordHash = hasher.HashPassword(null, "admin")
+        };
+
+        builder.HasData(admin);
+    }
+}
