+++ src/services/Identity/Identity.Infrastructure/DbContext/DbConfiguration/IdentityRoleConfiguration.cs
@@ -1,0 +1,19 @@
+using Microsoft.AspNetCore.Identity;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+namespace Identity.Infrastructure.DbConfiguration;
+
+public class IdentityRoleConfiguration : IEntityTypeConfiguration<IdentityRole<int>>
+{
+    public void Configure(EntityTypeBuilder<IdentityRole<int>> builder)
+    {
+        builder.HasData(
+            new List<IdentityRole<int>>()
+            {
+                Domain.Constraints.Roles.UserRole,
+                Domain.Constraints.Roles.AdminRole
+            }
+        );
+    }
+}
