+++ src/services/Identity/Identity.Infrastructure/DbContext/DbConfiguration/IdentityUserRoleConfiguration.cs
@@ -1,0 +1,14 @@
+using Microsoft.AspNetCore.Identity;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+namespace Identity.Infrastructure.DbConfiguration
+{
+    public class IdentityUserRoleConfiguration : IEntityTypeConfiguration<IdentityUserRole<int>>
+    {
+        public void Configure(EntityTypeBuilder<IdentityUserRole<int>> builder)
+        {
+            builder.HasData(new IdentityUserRole<int>() { RoleId = 2, UserId = 1 });
+        }
+    }
+}
